requirements.txt
BeautifulSoup==3.2.1
Flask==0.10.1
Flask-Migrate==1.3.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-RESTful==0.3.2
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
SQLAlchemy==0.9.8
Werkzeug==0.9.6
alembic==0.7.1
argparse==1.2.1
gunicorn==19.1.1
gevent==1.0.1
itsdangerous==0.24
psycopg2==2.5.4
requests==2.4.3
wsgiref==0.1.2
facebook-sdk==0.4.0
